requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
moviepy==1.0.3
edge-tts==6.1.12
requests==2.32.3
pillow==10.4.0
psutil>=5.9
accelerate==0.33.0
tiktoken==0.7.0
protobuf==5.27.3
